+++ packages/react-router/modules/__tests__/withRouter-test.js
@@ -64,4 +64,49 @@ describe('withRouter', () => {
       ), node)
     })
   })
+
+  it('exposes the wrapped component as WrappedComponent', () => {
+    const Component = () => <div/>
+    const decorated = withRouter(Component)
+    expect(decorated.WrappedComponent).toBe(Component)
+  })
+
+  it('exposes the instance of the wrapped component via wrappedComponentRef', () => {
+    class WrappedComponent extends React.Component {
+      render() {
+        return null
+      }
+    }
+    const Component = withRouter(WrappedComponent)
+
+    let ref
+    ReactDOM.render((
+      <MemoryRouter initialEntries={[ '/bubblegum' ]}>
+        <Route path="/bubblegum" render={() => (
+          <Component wrappedComponentRef={r => ref = r}/>
+        )}/>
+      </MemoryRouter>
+    ), node)
+
+    expect(ref).toBeA(WrappedComponent)
+  })
+
+  it('hoists non-react statics from the wrapped component', () => {
+    class Component extends React.Component {
+      static foo() {
+        return 'bar'
+      }
+
+      render() {
+        return null
+      }
+    }
+    Component.hello = 'world'
+
+    const decorated = withRouter(Component)
+
+    expect(decorated.hello).toBe('world')
+    expect(decorated.foo).toBeA('function')
+    expect(decorated.foo()).toBe('bar')
+  })
 })